Move ContextBase into core
