+++ scripts/templates/context.cpp
@@ -1,4 +1,4 @@
-#include <context.hpp>
+#include <core/context_base.hpp>
 #include <core/devices_manager.hpp>
 #include <devices_table.hpp>
